Enable descriptor set generation
Also:
 * Move extensions from the `Plugin.kt` to specific files.
